feat: 实现基于用户行为的推荐系统
- 新增UserBehaviorWeightDto用于记录用户综合行为权重
- 修改InteractionMapper添加行为权重查询方法
- 重构RecommendServiceImpl实现多策略推荐算法:
  1. 协同过滤推荐(基于浏览、点赞、收藏、评分)
  2. 分类偏好推荐
  3. 热门推荐
  4. 随机推荐(兜底)
- 更新项目文档说明推荐系统架构

+++ food-share-api/src/main/java/cn/kmbeast/pojo/dto/query/extend/RatingDto.java
@@ -6,6 +6,7 @@ import lombok.NoArgsConstructor;
 
 /**
  * 评分的Dto类
+ * 用于推荐算法中的用户-物品评分数据
  */
 @Data
 @AllArgsConstructor
@@ -20,7 +21,7 @@ public class RatingDto {
      */
     private Integer itemId;
     /**
-     * 评分
+     * 评分（支持浮点数，用于综合行为权重）
      */
-    private Integer score;
+    private Float score;
 }
